fix: core→clean dipendenza risolta, ruff script fixed
- Spostati read_sql_utils.py e read_csv_normalized.py in core/
  (erano l'ultima dipendenza core→clean ancora presente)
- core/duckdb_read.py ora importa solo da core/
- Fix ruff: variabili inutilizzate, import morto, f-string in script
- Shim backward compat in clean/ per tutti i moduli spostati
