+++ toolkit/core/duckdb_read.py
@@ -6,9 +6,9 @@ from pathlib import Path
 from typing import Any
 
 import duckdb
-from toolkit.clean.read_csv_normalized import _execute_normalized_csv_read
+from toolkit.core.read_csv_normalized import _execute_normalized_csv_read
 from toolkit.core.read_excel import _execute_excel_read
-from toolkit.clean.read_sql_utils import (
+from toolkit.core.read_sql_utils import (
     _parse_column_value,
     csv_trim_projection,
 )
